Bug #2285: Do not run code in swapstate_check.php when there is no Squid cache

+++ config/squid/swapstate_check.php
@@ -30,19 +30,23 @@ require_once('config.inc');
 require_once('util.inc');
 
 $settings = $config['installedpackages']['squidcache']['config'][0];
-$cachedir =($settings['harddisk_cache_location'] ? $settings['harddisk_cache_location'] : '/var/squid/cache');
-$swapstate = $cachedir . '/swap.state';
-$disktotal = disk_total_space(dirname($cachedir));
-$diskfree = disk_free_space(dirname($cachedir));
-$diskusedpct = round((($disktotal - $diskfree) / $disktotal) * 100);
-$swapstate_size = filesize($swapstate);
-$swapstate_pct = round(($swapstate_size / $disktotal) * 100);
+// Only check the cache if Squid is actually caching.
+// If there is no cache then quietly do nothing.
+if ($settings['harddisk_cache_system'] != "null"){
+	$cachedir =($settings['harddisk_cache_location'] ? $settings['harddisk_cache_location'] : '/var/squid/cache');
+	$swapstate = $cachedir . '/swap.state';
+	$disktotal = disk_total_space(dirname($cachedir));
+	$diskfree = disk_free_space(dirname($cachedir));
+	$diskusedpct = round((($disktotal - $diskfree) / $disktotal) * 100);
+	$swapstate_size = filesize($swapstate);
+	$swapstate_pct = round(($swapstate_size / $disktotal) * 100);
 
-// If the swap.state file is taking up more than 75% disk space, 
-//	or the drive is 90% full and swap.state is larger than 1GB, 
-//	kill it and initiate a rotate to write a fresh copy.
-if (($swapstate_pct > 75) || (($diskusedpct > 90) && ($swapstate_size > 1024*1024*1024))) {
-	mwexec_bg("/bin/rm $swapstate; /usr/local/sbin/squid -k rotate");
-	log_error(gettext(sprintf("Squid swap.state file exceeded size limits. Removing and rotating. File was %d bytes, %d%% of total disk space.", $swapstate_size, $swapstate_pct)));
+	// If the swap.state file is taking up more than 75% disk space, 
+	//	or the drive is 90% full and swap.state is larger than 1GB, 
+	//	kill it and initiate a rotate to write a fresh copy.
+	if (($swapstate_pct > 75) || (($diskusedpct > 90) && ($swapstate_size > 1024*1024*1024))) {
+		mwexec_bg("/bin/rm $swapstate; /usr/local/sbin/squid -k rotate");
+		log_error(gettext(sprintf("Squid swap.state file exceeded size limits. Removing and rotating. File was %d bytes, %d%% of total disk space.", $swapstate_size, $swapstate_pct)));
+	}
 }
 ?>
